Include requirement-setuptools.txt in manifest
When attempting to `pip install ckan` it fails as the setup.py requires
the requirement-setuptools.txt to be present, and it is not.  This
change makes sure that the pip install will not fail due to the absence
of this file.

This PR does not include the requirements.txt as this would likely
require further discussion about how requirements are loaded.

Fixes #4271