Replace HTTPIE examples with CURL examples
As per #2293 we replace the httpie examples with the equivalent in curl.
It appears that the examples were using http and therefore were getting
a 302 to the https url, and as we all know to our pain, POST data
doesn't follow redirects and so the examples weren't always working.
Have changes some occurrences of http://demo.ckan.org to
https://demo.ckan.org